requirements.txt
pyodbc==4.0.39
python-dotenv==1.0.0
tabulate==0.9.0
flask==2.3.3
flask-wtf==1.1.1